refactor: asset 파일 정리
