--- src/mocks/courseMock.json
@@ -1,110 +1,0 @@
-[
-  {
-    "courseCode": 10,
-    "prosessor": "송명진",
-    "courseName": "사고와 표현(mock 데이터)",
-    "courseLink": "https://www.naver.com",
-    "courseIdentifier": "A0001_0001",
-    "lectureRoom": "3306(멀티미디어실 New)",
-    "notices": [
-      {
-        "link": "https://www.naver.com",
-        "content": "근데 미헬스는 그 네덜란드의 토탈풋볼을 선보이기 위해 선수들을 아주 갈아마셔버렸다고 알려져있죠. 이미 그런 축구에 익숙해있던 아약스 선수들을 제외한 나머지 선수들은 그런 걸 이해도 못했고 몸이 바로바로 반응이 안 됐으니까요. 그래서 훈련을 더 시켰습니다. 우리가 생각하는 것 이상으로... 안 되면 될 때까지 해라. 이런 느낌으로 말이죠. 그래서 미헬스의 별명은 장군이었죠. (우리나라로 따지면 독불장군에 가까운 장군이겠죠?) 독단적이고 자기 말을 따르지 않는 선수들은 철저하게 배제하는 독재자란 평가를 받기도 했었구요.",
-        "index": "1",
-        "title": "공지사항 -1 ",
-        "date": "2025-01-09",
-        "files": [
-          {
-            "fileName": "[양식1]보고서 제출",
-            "fileLink": "https://www.naver.com"
-          },
-          {
-            "fileName": "[양식2]개요서 제출",
-            "fileLink": "https://www.naver.com"
-          }
-        ]
-      },
-      {
-        "link": "https://www.naver.com",
-        "content": "옛날로 돌아가보면 사실 74 월드컵 네덜란드는 지금의 시선으로 보면 되게 형편없습니다. 보다가 계속 꺼버리고 결국엔 대부분의 경기들을 풀 타임 시청을 못하긴 했는데 (아리고 사키의 밀란까진 어떻게 됐는데 그 이전은 풀 타임 시청이 안 되더라구요. 원래 다시 보기라는 거 자체를 안 좋아하긴 하는데 옛날 경기 다 보고나서 다시 보기는 정말 할 게 못 된다는 걸 느끼고 그 후로 지나간 경기들이나 옛날 경기들은 안 찾아봅니다.) 모두가 미친듯이 뛰어다니긴 하는데 되게 무질서하고 중구난방이었습니다. 근데 당시에 충격적이었던 건 그거죠. 축구란 스포츠는 긴 거리를 돌파하고 상대 수비수들을 박스 근처에서 현란하게 제끼면서 골키퍼를 넘어서는 그런 스포츠였는데 다른 의미로 접근한 거였으니까요.",
-        "index": "2",
-        "title": "공지사항 -2 ",
-        "date": "2025-01-10",
-        "files": [
-          {
-            "fileName": "[양식1]보고서 제출",
-            "fileLink": "https://www.naver.com"
-          },
-          {
-            "fileName": "[양식2]개요서 제출",
-            "fileLink": "https://www.naver.com"
-          }
-        ]
-      }
-    ],
-    "activities": [
-      {
-        "week": 1,
-        "lectures": [
-          {
-            "activityCode": 1,
-            "activityName": "1주차 플립드러닝-1",
-            "date": "9월20일-9월26일",
-            "activityType": "video",
-            "activityLink": "https://www.naver.com"
-          },
-          {
-            "activityCode": 2,
-            "activityName": "1주차 플립드러닝-2",
-            "date": "9월20일-9월26일",
-            "activityType": "video",
-            "activityLink": "https://www.naver.com"
-          }
-        ],
-        "assignment": []
-      },
-      {
-        "week": 2,
-        "lectures": [
-          {
-            "activityCode": 4,
-            "activityName": "2주차 플립드러닝-1",
-            "date": "9월30일-10월7일",
-            "activityType": "video",
-            "activityLink": "https://www.naver.com"
-          }
-        ],
-        "assignment": [
-          {
-            "activityCode": 5,
-            "activityName": "2주차 과제",
-            "date": "10월8일-10월13일",
-            "activityType": "assignment",
-            "activityLink": "https://www.naver.com"
-          }
-        ]
-      },
-      {
-        "week": 3,
-        "lectures": [
-          {
-            "activityCode": 6,
-            "activityName": "2주차 플립드러닝-1",
-            "date": "10월8일-10월13일",
-            "activityType": "video",
-            "activityLink": "https://www.naver.com"
-          }
-        ],
-        "assignment": [
-          {
-            "activityCode": 7,
-            "activityName": "2주차 과제",
-            "date": "9월20일-9월26일",
-            "activityType": "assignment",
-            "activityLink": "https://www.naver.com"
-          }
-        ]
-      }
-    ]
-  }
-]
