--- src/mocks/courseMocks.json
@@ -1,198 +1,0 @@
-{
-  "courses": [
-    {
-      "courseCode": "1001",
-      "courseName": "사고와표현",
-      "courseLink": "https://www.naver.com",
-      "courseIdentifier": "A0001_0001",
-      "professor": "송명진",
-      "lectureRoom": "3306(멀티미디어실 New)",
-      "notices": [
-        {
-          "link": "https://www.naver.com",
-          "content": "공지사항 내용입니다. \n공지사항 내용입니다. \n공지사항 내용입니다. \n",
-          "index": "1",
-          "title": "공지사항 -1 ",
-          "date": "2025-01-09",
-          "files": [
-            {
-              "fileName": "[양식1]보고서 제출",
-              "fileLink": "https://www.naver.com"
-            },
-            {
-              "fileName": "[양식2]개요서 제출",
-              "fileLink": "https://www.naver.com"
-            }
-          ]
-        },
-        {
-          "link": "https://www.naver.com",
-          "content": "공지사항 내용입니다. \n공지사항 내용입니다. \n공지사항 내용입니다. \n",
-          "index": "2",
-          "title": "공지사항 -2 ",
-          "date": "2025-01-10",
-          "files": [
-            {
-              "fileName": "[양식1]보고서 제출",
-              "fileLink": "https://www.naver.com"
-            },
-            {
-              "fileName": "[양식2]개요서 제출",
-              "fileLink": "https://www.naver.com"
-            }
-          ]
-        }
-      ],
-      "activities": [
-        [
-          {
-            "week": 1,
-            "available": "true",
-            "activityName": "1주차 강의 자료",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000001",
-            "activityType": "file"
-          },
-          {
-            "week": 1,
-            "available": "true",
-            "activityName": "1주차 영상 강의",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000002",
-            "activityType": "video",
-            "lectureDeadline": "2025-01-10 00:00:00 ~ 2025-01-15 23:59:59",
-            "lectureTime": "22:57"
-          }
-        ],
-        [
-          {
-            "week": 2,
-            "available": "true",
-            "activityName": "2주차 강의 자료",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000003",
-            "activityType": "file"
-          },
-          {
-            "week": 2,
-            "available": "true",
-            "activityName": "2주차 영상 강의",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000004",
-            "activityType": "video",
-            "lectureDeadline": "2025-01-10 00:00:00 ~ 2025-01-15 23:59:59",
-            "lectureTime": "22:57"
-          },
-          {
-            "week": 2,
-            "available": "true",
-            "activityName": "2주차 과제",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000005",
-            "activityType": "assigment",
-            "gradingStatus": "채점되지 않음.",
-            "deadline": "2025-01-10 00:00:00 ~ 2025-01-15 23:59:59",
-            "description": "2주차 과제는 pdf로 제출 바랍니다."
-          }
-        ]
-      ]
-    },
-    {
-      "courseCode": "1002",
-      "courseName": "웹프로그래밍",
-      "courseLink": "https://www.naver.com",
-      "courseIdentifier": "A0001_0002",
-      "professor": "정경용",
-      "lectureRoom": "8001",
-      "notices": [
-        {
-          "link": "https://www.naver.com",
-          "content": "공지사항 내용입니다. \n공지사항 내용입니다. \n공지사항 내용입니다. \n",
-          "index": "1",
-          "title": "공지사항 -1 ",
-          "date": "2025-01-09",
-          "files": [
-            {
-              "fileName": "[양식1]보고서 제출",
-              "fileLink": "https://www.naver.com"
-            },
-            {
-              "fileName": "[양식2]개요서 제출",
-              "fileLink": "https://www.naver.com"
-            }
-          ]
-        },
-        {
-          "link": "https://www.naver.com",
-          "content": "공지사항 내용입니다. \n공지사항 내용입니다. \n공지사항 내용입니다. \n",
-          "index": "2",
-          "title": "공지사항 -2 ",
-          "date": "2025-01-10",
-          "files": [
-            {
-              "fileName": "[양식1]보고서 제출",
-              "fileLink": "https://www.naver.com"
-            },
-            {
-              "fileName": "[양식2]개요서 제출",
-              "fileLink": "https://www.naver.com"
-            }
-          ]
-        }
-      ],
-      "activities": [
-        [
-          {
-            "week": 1,
-            "available": "true",
-            "activityName": "웹프 1주차 강의 자료",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000001",
-            "activityType": "file"
-          },
-          {
-            "week": 1,
-            "available": "true",
-            "activityName": "1주차 영상 강의",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000002",
-            "activityType": "video",
-            "lectureDeadline": "2025-01-10 00:00:00 ~ 2025-01-15 23:59:59",
-            "lectureTime": "22:57"
-          }
-        ],
-        [
-          {
-            "week": 2,
-            "available": "true",
-            "activityName": "2주차 강의 자료",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000003",
-            "activityType": "file"
-          },
-          {
-            "week": 2,
-            "available": "true",
-            "activityName": "2주차 영상 강의",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000004",
-            "activityType": "video",
-            "lectureDeadline": "2025-01-10 00:00:00 ~ 2025-01-15 23:59:59",
-            "lectureTime": "22:57"
-          },
-          {
-            "week": 2,
-            "available": "true",
-            "activityName": "2주차 과제",
-            "activityLink": "https://www.naver.com",
-            "activityCode": "000005",
-            "activityType": "assigment",
-            "gradingStatus": "채점되지 않음.",
-            "deadline": "2025-01-10 00:00:00 ~ 2025-01-15 23:59:59",
-            "description": "2주차 과제는 pdf로 제출 바랍니다."
-          }
-        ]
-      ]
-    }
-  ]
-}
